proposal/serializers.py
from rest_framework import serializers
from .models import Proposal
from eav.models import Attribute


class AttributeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Attribute
        fields = ('name', 'slug',)


class ProposalSerializer(serializers.ModelSerializer):
    class Meta:
        model = Proposal
        fields = ('id', 'name', 'document', 'status', 'last_updated')
