requirements.txt
Django==4.2.4
djangorestframework==3.14.0
psycopg2-binary==2.9.7
dj-database-url==2.1.0
celery==5.2.7
redis==4.5.5
flake8==6.1.0
autopep8==2.0.3
django-eav2==1.4.0
coverage==7.3.0
requests==2.31.0
django-cors-headers==4.2.0